+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The types of changes are:
 ### Fixed
 
 - Datamap export mitigation for deleted taxonomy elements referenced by declarations [#3214](https://github.com/ethyca/fides/pull/3214)
+- Ensure inactive custom fields are not returned for datamap response [#3223](https://github.com/ethyca/fides/pull/3223)
 
 ## [2.12.0](https://github.com/ethyca/fides/compare/2.11.0...2.12.0)